requirements.txt
# Core stack
Django>=5.0,<6.0
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<6.0

# Database + async
psycopg2-binary>=2.9,<3.0
celery>=5.3,<5.4
redis>=5.0,<6.0

# LLM
httpx>=0.27,<0.28
pydantic>=2.7,<3.0
python-dotenv>=1.0,<2.0
structlog>=24.1,<25.0

# test/
pytest>=7.4,<9.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory_boy>=3.3,<4.0
black>=24.3,<25.0
ruff>=0.5,<0.6
